Fix downloader for windows

+++ download.cr
@@ -5,13 +5,13 @@ def url_for_libui_ng_nightly(file_name)
   "https://nightly.link/kojix2/libui-ng/workflows/pre-build/pre-build/#{file_name}"
 end
 
-def download_libui_ng_nightly(libname, lib_path, file_name)
+def download_libui_ng_nightly(lib_path, file_name)
   url = url_for_libui_ng_nightly(file_name)
-  download_from_url(libname, lib_path, file_name, url)
+  download_from_url(lib_path, file_name, url)
 end
 
-def download_from_url(libname, lib_path, file_name, url)
-  STDERR.puts "Downloading #{libname} from #{url}"
+def download_from_url(lib_path, file_name, url)
+  STDERR.puts "Downloading #{lib_path} from #{url}"
 
   Crest.get(url) do |response|
     File.open(file_name, "wb") do |file|
@@ -22,9 +22,9 @@ def download_from_url(libname, lib_path, file_name, url)
   if file_name.ends_with?(".zip")
     Compress::Zip::File.open(file_name) do |zip_file|
       zip_file.entries.each do |entry|
-        if entry.filename == lib_path
+        if lib_path.includes?(entry.filename)
           entry.open do |io|
-            File.open(libname, "wb") do |file|
+            File.open(File.basename(entry.filename), "wb") do |file|
               IO.copy(io, file)
             end
           end
@@ -32,26 +32,23 @@ def download_from_url(libname, lib_path, file_name, url)
       end
     end
   end
-
-  File.delete(file_name)
+ensure
+  File.delete(file_name) if File.exists?(file_name)
 end
 
 {% if flag?(:darwin) %}
   download_libui_ng_nightly(
-    "libui.dylib",
-    "builddir/meson-out/libui.dylib",
+    ["builddir/meson-out/libui.dylib"],
     "macOS-x64-shared-release.zip"
   )
 {% elsif flag?(:linux) %}
   download_libui_ng_nightly(
-    "libui.so",
-    "builddir/meson-out/libui.so",
+    ["builddir/meson-out/libui.so"],
     "Ubuntu-x64-shared-release.zip"
   )
 {% elsif flag?(:windows) %}
   download_libui_ng_nightly(
-    "libui.dll",
-    "builddir/meson-out/libui.dll",
+    ["builddir/meson-out/libui.dll", "builddir/meson-out/libui.lib"],
     "Win-x64-shared-release.zip"
   )
 {% end %}
